Renamed gc_command.rs to clean_command.rs.
