+++ src/main.rs
@@ -24,6 +24,7 @@ mod attributes;
 mod background_executer;
 mod backup_command;
 mod backup_store;
+mod clean_command;
 mod commons;
 mod config;
 mod entry;
@@ -31,7 +32,6 @@ mod entry_saver;
 mod error;
 mod file_path_producer;
 mod garbage_collector;
-mod gc_command;
 mod get_command;
 mod init_command;
 mod object_adder;
@@ -45,7 +45,7 @@ use std::process::ExitCode;
 use std::time::SystemTime;
 
 use crate::backup_command::BackupCommand;
-use crate::gc_command::GcCommand;
+use crate::clean_command::GcCommand;
 use crate::get_command::GetCommand;
 use crate::init_command::InitCommand;
 use crate::remove_backup_command::RemoveBackupCommand;
